Step 3 Store and retrieve orders
Service is able to now store orders that a user submits. Two endpoints
have been added: `/get-order` and `/get-all-orders`. By submitting a
POST request to the `/get-order` endpoint with a payload containing an
order_id (see README for more info), the service responds with an order
summary if the order was found. Sending a GET request to the
`/get-all-orders` endpoint will return turn a list of all orders that
have been successfully submitted.